requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
billiard==4.2.4
cairocffi==1.7.1
CairoSVG==2.8.2
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.13.1
cron_descriptor==2.0.6
cryptography==46.0.3
cssselect2==0.8.0
# Editable install with no version control (csu-tracker==0.1.0)
-e /Users/admin/Downloads/CSU_APP
decorator==5.2.1
defusedxml==0.7.1
dj-database-url==3.1.0
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
frozenlist==1.8.0
gunicorn==23.0.0
http_ece==1.2.1
idna==3.11
iniconfig==2.3.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.6.2
matplotlib-inline==0.2.1
multidict==6.7.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg==3.3.2
psycopg-binary==3.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
py-vapid==1.9.4
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
pywebpush==2.2.0
redis==6.4.0
requests==2.32.5
ruff==0.14.13
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
tinycss2==1.5.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
webencodings==0.5.1
whitenoise==6.11.0
yarl==1.22.0
